# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ootils Core API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/ingest/bom": {
      "post": {
        "tags": [
          "bom"
        ],
        "summary": "Import BOM",
        "description": "Import a complete Bill of Materials for an item. Automatically computes Low-Level Codes (LLC).",
        "operationId": "ingest_bom_v1_ingest_bom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBOMRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBOMResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/bom/{parent_external_id}": {
      "get": {
        "tags": [
          "bom"
        ],
        "summary": "Get BOM",
        "description": "Return the active BOM for an item with its components and LLC.",
        "operationId": "get_bom_v1_bom__parent_external_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parent_external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Parent External Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bom/explode": {
      "post": {
        "tags": [
          "bom"
        ],
        "summary": "MRP explosion",
        "description": "Compute gross and net component requirements for a given quantity (multi-level explosion, netting against available stock).",
        "operationId": "explode_bom_v1_bom_explode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BOMComponentInput": {
        "properties": {
          "component_external_id": {
            "type": "string",
            "title": "Component External Id",
            "description": "Component external_id."
          },
          "quantity_per": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity Per",
            "description": "Component quantity per 1 unit of parent (> 0)."
          },
          "uom": {
            "type": "string",
            "title": "Uom",
            "description": "Component unit of measure.",
            "default": "EA"
          },
          "scrap_factor": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Scrap Factor",
            "description": "Scrap rate [0.0–1.0). E.g. 0.05 = 5% waste.",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "component_external_id",
          "quantity_per"
        ],
        "title": "BOMComponentInput"
      },
      "BOMComponentOutput": {
        "properties": {
          "component_external_id": {
            "type": "string",
            "title": "Component External Id"
          },
          "component_item_id": {
            "type": "string",
            "title": "Component Item Id"
          },
          "quantity_per": {
            "type": "number",
            "title": "Quantity Per"
          },
          "uom": {
            "type": "string",
            "title": "Uom"
          },
          "scrap_factor": {
            "type": "number",
            "title": "Scrap Factor"
          },
          "llc": {
            "type": "integer",
            "title": "Llc"
          }
        },
        "type": "object",
        "required": [
          "component_external_id",
          "component_item_id",
          "quantity_per",
          "uom",
          "scrap_factor",
          "llc"
        ],
        "title": "BOMComponentOutput"
      },
      "BOMResponse": {
        "properties": {
          "parent_external_id": {
            "type": "string",
            "title": "Parent External Id"
          },
          "parent_item_id": {
            "type": "string",
            "title": "Parent Item Id"
          },
          "bom_version": {
            "type": "string",
            "title": "Bom Version"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/BOMComponentOutput"
            },
            "type": "array",
            "title": "Components"
          }
        },
        "type": "object",
        "required": [
          "parent_external_id",
          "parent_item_id",
          "bom_version",
          "effective_from",
          "components"
        ],
        "title": "BOMResponse"
      },
...
      "ExplodeLineOutput": {
        "properties": {
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "component_external_id": {
            "type": "string",
            "title": "Component External Id"
          },
          "component_item_id": {
            "type": "string",
            "title": "Component Item Id"
          },
          "gross_requirement": {
            "type": "number",
            "title": "Gross Requirement"
          },
          "on_hand_qty": {
            "type": "number",
            "title": "On Hand Qty"
          },
          "net_requirement": {
            "type": "number",
            "title": "Net Requirement"
          },
          "has_shortage": {
            "type": "boolean",
            "title": "Has Shortage"
          }
        },
        "type": "object",
        "required": [
          "level",
          "component_external_id",
          "component_item_id",
          "gross_requirement",
          "on_hand_qty",
          "net_requirement",
          "has_shortage"
        ],
        "title": "ExplodeLineOutput"
      },
      "ExplodeRequest": {
        "properties": {
          "item_external_id": {
            "type": "string",
            "title": "Item External Id",
            "description": "Item to explode."
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Quantity to produce (> 0)."
          },
          "location_external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location External Id",
            "description": "Production site. Optional."
          },
          "explosion_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explosion Date",
            "description": "Reference date for explosion. Defaults to today."
          },
          "levels": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Levels",
            "description": "Maximum explosion depth [1–20]. Default = 10.",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "item_external_id",
          "quantity"
        ],
        "title": "ExplodeRequest"
      },
      "ExplodeResponse": {
        "properties": {
          "parent_external_id": {
            "type": "string",
            "title": "Parent External Id"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "explosion": {
            "items": {
              "$ref": "#/components/schemas/ExplodeLineOutput"
            },
            "type": "array",
            "title": "Explosion"
          },
          "total_components": {
            "type": "integer",
            "title": "Total Components"
          },
          "components_with_shortage": {
            "type": "integer",
            "title": "Components With Shortage"
          }
        },
        "type": "object",
        "required": [
          "parent_external_id",
          "quantity",
          "explosion",
          "total_components",
          "components_with_shortage"
        ],
        "title": "ExplodeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestBOMRequest": {
        "properties": {
          "parent_external_id": {
            "type": "string",
            "title": "Parent External Id",
            "description": "Parent item external_id (manufactured item)."
          },
          "bom_version": {
            "type": "string",
            "title": "Bom Version",
            "description": "BOM version (e.g. '1.0', '2.1').",
            "default": "1.0"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From",
            "description": "BOM effective date."
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/BOMComponentInput"
            },
            "type": "array",
            "title": "Components",
            "description": "BOM component list."
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If true, validation only — no DB writes.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "parent_external_id",
          "components"
        ],
        "title": "IngestBOMRequest"
      },
      "IngestBOMResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bom_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bom Id"
          },
          "parent_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Item Id"
          },
          "components_imported": {
            "type": "integer",
            "title": "Components Imported"
          },
          "llc_updated": {
            "type": "integer",
            "title": "Llc Updated"
          }
        },
        "type": "object",
        "required": [
          "status",
          "components_imported",
          "llc_updated"
        ],
        "title": "IngestBOMResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}